Move FFI::StructLayout to its own file
